requirements.txt
fastapi~=0.103.1
pydantic~=2.3.0
requests~=2.31.0
beautifulsoup4~=4.12.2
selenium~=4.12.0
PyYAML~=6.0.1
uvicorn~=0.23.2